requirements-dev.txt
# Development dependencies for PyBulletFleet

# Code formatting
black>=23.0.0
isort>=5.12.0

# Linting
flake8>=6.0.0

# Type checking
mypy>=1.0.0
types-PyYAML>=6.0.0
types-setuptools>=65.0.0

# Pre-commit hooks
pre-commit>=3.0.0

# Testing (for future use)
pytest>=7.0.0
pytest-cov>=4.0.0

# Runtime dependencies (in case requirements.txt is not installed)
pybullet>=3.0.0
numpy>=1.20.0
pyyaml>=5.4.0
